Handle postgres not running on instance destroy
Currently, if postgres (for whatever reason) isn't running when we try
to destroy an instance, we'll exit early because pg_ctl will return a
nonzero exit code.

If postgres isn't running then we don't really care, since we're about
to destroy the directoy, so we add an or condition at the end of the
pg_ctl call so that `set -e` doesn't cause us to exit early.