Implement serverside oauth flow
In this flow, the client sends the user to server.tld/authenticate, which
redirects to the oauth connection page, passing a unique state parameter.
At the same time, the client makes a request to the server to create an
access token, with the same state. The server holds this connection open.
When the oauth connection completes, the server will receive the callback
and return the access token to the client via the open request.

This way, we can do oauth without storing any sensitive credentials in the
client.