CircleCI integration tests w/o Chef
Remove the dependence on the draupnir Chef cookbook to run integration
tests, vastly simplifying and increasing the speed of running tests from
a local development machine.

This is achieved by using plain dockerfiles and a wrapper in the rspec
test suite to provision a new draupnir instance each time the tests are
run. We also introduce a CircleCI workflow that splits compilation of
the go binary, unit testing of said binary and running rspec integration
tests into separate dependent jobs.