Use refresh tokens for oauth
So that we can reduce the frequency with which users have to re-authenticate,
send oauth refresh tokens to the server. These can be used to generate
automatically-refreshing access tokens.

We store the full oauth2.Token struct on the client to give us the freedom to
change this approach in the future.