Switch config format to TOML and add default database config
Use TOML as the format for the .draupnir config file. It's easier to
read and edit than JSON.

Add a config item "database", which is the default database the cli will
use when rendering environment variables for an instance. If not set, it
will fall back to the PGDATABASE, and if that's not set it will default
to "postgres".

Allow the database and domain config items to be set via the command
line, so you don't need to manually edit the config file.