Semver match on Draupnir-Version header
Only reject requests that come from clients that are newer than the
server, or are running a different major version. This should be totally
fine as per semver rules, and will provide more flexibility to upgrade
the server without breaking clients.

The client also now doesn't inject an error if the server version
mismatches, and leaves the check up to the server. The API will return
an error that would otherwise be rendered, so we don't need to duplicate
the check in the client.