Makefile tweaks
- Add client to PHONY targets
- Allow building of binary without running tests

The first is required to build the client without manually cleaning, and
the second means it's much faster to generate a development binary. We
have CI that runs on every build and we shouldn't be deploying anything
that isn't green, so forcing a developer to locally run tests for every
binary isn't beneficial.