Optionally unpack tar backup
draupnir-finalise-image will currently fail if the upload is not in tar
format. This change allows a user to do their own custom upload of a
draupnir data directory, with whatever preparation step they wish to
perform, by only unpacking a tar file if we find one present.

This becomes really useful for people who wish to do the backup recovery
on the same machine as draupnir, as they can unpack the database and
recover from WAL files all inside the image upload directory.