Improve finalisation performance
Running large anonymisation scripts writes loads of data to disk. One
thing that can significantly impact performance is the amount of WAL we
write to disk, and whether Postgres calls fsync to guarantee data
integrity after each write.

We can make finalisation much faster by disabling fsync (as we can
easily recreate a draupnir image from a backup, so we're not worried
about crash-safety here) and reducing the amount of data we write to our
WAL files.

We were previously using hot_standby WAL level, which meant we wrote all
the data required to power live read queries on a standby, which is
significantly more than the minimal/warm-standby level. We can change to
minimal for the anonymisation work, but should change back to
hot_standby for the live instances to ensure they better mirror
production performance.