Vacuum databases after running finalisation
The finalisation scripts usually rewrite a lot of data in order to
anonymise the production backup. This leaves the database in a state
where vacuuming is required, as we've crossed autovacuum thresholds.

When you have many instances of Postgres all vacuuming on the same btrfs
filesystem, you amplify the writes by the number of instances you run,
and result in generating huge amounts of data exclusive to each
snapshot.

This change has draupnir vacuum all the tables in the database, in
parallel, to ensure the fresh instances are far from requiring a vacuum.
We use psql and xargs to run the vacuum in parallel as vacuumdb only
supports the --jobs flag in 9.5 and above (which we don't run, sadly).