Fix logging in executor package
There are two changes here:

1. Use the correct context key in order to retrieve the logger. Prior
to this the no-op logger was being assigned, and hence we weren't
receiving any log output from these operations.

2. Extract the logic to run and log the output of a command to a helper
function. This ensures we always capture the error code/status, stdout
and stderr from the execution of our commands, which is a massive
improvement when trying to debug this part of the system.