Fix error-passing in middlewares
Several of the middlewares were not passing on errors that had occurred
either in the previous middleware or the main handler function itself.
This caused the server to always return 'successful' status codes even
in case of a major error, and to not log the error message of the root
cause.

This should also prevent the 'EOF' errors seen in the Draupnir client,
as the server side now always correctly sends an error response.