Allow configuration of an instance hostname
This allows us to split the API hostname and the hostname used to
connect to Postgres instances, because we can't route Postgres traffic
through the same load balancer as the API requests.

By having the server-side set the hostname that the client should
connect to, we lay the foundations for scaling Draupnir to more than one
node too.