Decrease the potential for port collisions
Currently there is a high chance of collisions when booting a new
Draupnir instance. As per the birthday problem formula, if there are 10
instances running, with a port range of 1000 ports, then we've got a 10%
chance of colliding. When this happens the instance fails to start, and
the user is returned an unhelpful error about an internal server error.

To resolve this, two changes have been made:
1. The random number generator is seeded with nanoseconds, rather than
   just seconds, which helps us in the case of two requests that arrive
   within the same second.
2. The list of existing instances is consulted, and if the
   randomly-chosen port is already assigned to an instance then another is
   picked instead.