Verify instance auth settings upon creation
The security of instances currently depends on the fact that an image
has been finalised with a `hostssl ... trust clientcert=1` line in the
`pg_hba.conf`.
Without this line present, and in the correct format, it'd be possible
for anyone to connect to an instance without providing credentials.

As an additional verification step, add checks at instance-creation time
that ensure that it's not possible to connect to an instance without
TLS, or without providing a client certificate. While these checks are
being performed, the instance is listening on the local interface only,
and therefore is inaccessible to external users.

If these checks fails then the instance is shut down and the
provisioning process returns an error, which will be sent back to the
client and trigger a Sentry.