Add whitelister component
This provides dynamic IP address whitelisting for instances, using
iptables rules, in order to add an extra layer of protection.

This can be enabled by setting `enable_ip_whitelisting = true` in the
server configuration. No client changes are required: any API call to
create an instance or retrieve its details will result in the user's IP
address being stored in a whitelist and the iptables rules updated to
reflect this.

See the new section in `README.md` for further details of the design.