Add user IP address middleware
This is provides a user's 'real' IP address in the request context, as
well as including it in the request logger.

The implementation is a little complex as we need to check the
`X-Forwarded-For` HTTP header, but remove any entries from this that we
know to represent our own load balancers, in order to be able to
determine the actual source IP address, without the user being able to
spoof it.