Only allow instance connections from non-superuser
During the security review of the system concerns were raised about the
ability to connect to the Draupnir instances via the `postgres` user,
which has superuser privileges.
This is an issue because this user can use superuser-level functions to
do things like read file contents, and load extensions which may contain
malicious code.

To resolve this, make the following changes:
- Only allow connections from a `draupnir` user, which is added at
  image-finalisation time.
- Reassign all objects in all databases to the `draupnir` user, so that
  it has permissions to create, delete, alter etc all existing objects.
- Perform the instance verification checks with the `draupnir` user,
  rather than `postgres`.

Given that we now only allow connections from a single user, also change
the authentication in `pg_hba.conf` to enforce certificate
authentication, as opposed to trust but secured with certificate
authentication.