Improve configuration of listen addresses
In the current state there is an 'insecure' port, which is bound to
localhost only, and a 'secure' port that listens on all interfaces.

The secure port only accepts requests that have negotiated a connection
over TLS (using the certificate/key pair provided to Draupnir), whereas
the insecure port uses plain HTTP.

There's a few issues with this:

1. It's not possible to disable either the secure or insecure ports.
2. It's not possible to make the insecure server available on any
   interface other than localhost, even if you have firewall rules that
   mitigate that being a potential issue.
3. It's not possible to specify the address (and therefore interface)
   that the secure server listens, making it available to anything on
   the machine.

In our configuration we would like to expose an insecure address only,
as this will be locked down and fronted by a cloud load balancer.

Therefore change the 'port' settings to represent full listen addresses,
for maximum configurability, and allow either to be disabled by not
specifying them in the configuration.